+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v24.23.0
+
+- Added new cacheable plugin to allow dynamically updated variables to be written back to SSM Parameter Store/Secrets Manager. For more detail see `open-task-framework` documentation for version 24.22.0
+- Minor tweaks to SSM lookup plugin
+
 ## v24.19.0
 
 - Removed a stray `break` in S3 file listing that was preventing fetching more than the first 1000 records